this update including updated file handling. Using the path libary i now made it so that the setting, white list and device list files, are stored in the same file pathway no matter where u launch it. File pathways include: settings = settings.json active devices found list = active_device_list.json white list = white_list.json
Future improvements:
Want to add a better looking setting menu allowing for the user to view values/ json settings.
so that way i can view values and not have to change values to know for a fact what is there.
Will also be implementing deauth soon once i start using this program on Kali-Linux