Updated the settings module allowing users to see within the program values for all variables within the settings json file without having the go through each settings option.
Also updated the json file added monitor_mode_scans, which keeps track of the amount of scans completed
by monitor mode and will be having those same numbers show in the main menu within the welcome panel soon.

Also added N/A for host name and vendors that coudnt be found instead of unkown, also for playstation devices
for example if the vendor has sony in it i created a condition that checks for that and will change the host name
for that device to playstation if the host == "N/A" and if "sony" is in vendor
Will be adding more cool features to this soon and also add further redundency to make sure to false positives are created.